+++ lib/vpn_check.dart
@@ -1,2 +1,3 @@
 export 'src/vpn_check.dart';
+export 'src/builder.dart';
 export 'src/exception/vpn_exception.dart';